+++ Projectile.h
@@ -6,7 +6,7 @@ static bool ProjectileHitsBeing(Projectile* const, Being* const);
 void AddPCProjectileToArray(Projectiles_array* const, const SDL_FPoint* const, const float, const float, const Impact* const, const Uint8);
 void AddHProjectileToArray(Projectiles_array* const, const SDL_FPoint* const, const float, const float, const Impact* const);
 void AddSpecialProjectileToArray(Projectiles_array* const, const SDL_FPoint* const, const float, const float, const Uint8, const unsigned int);
-static void DestroyProjectileInArray(Projectiles_array* const, const Uint16);
+static void DestroyProjectileInArray(Projectiles_array* const, const unsigned int);
 static void MoveProjectile(Projectile* const);
 void UpdateProjectiles(Game_data* const);
 static bool ProjectileHitsPlayer(Projectile* const, Game_data* const, const unsigned int);
